requirements.txt
# pip freeze:

aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pycodestyle==2.6.0
pylint==2.6.0
python-engineio==3.13.2
python-socketio==4.6.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1